Obs 씬 제작 및 버튼 제작 및 프리펩 생성
장애물 담당자는 경옥이가 고기를 사야된다고 생각해요
