+++ DeathRun/Assets/Plane.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e919158d94136614ea3ed1a22c3e34a8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
