+++ DeathRun/Assets/Prefab/OBS.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 46d8dea1459c2f142a1aa326f6f0946d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
