+++ DeathRun/Assets/Prefab/OBS/Obs_Button.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 64b2ea974f7279344871e8e96a3da95a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
