+++ DeathRun/Assets/Scenes/Obs_test.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 877f40dfa5d9e1b478674b239fbffe93
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
