+++ DeathRun/Assets/Scripts/Obs_scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e6335bdd6e5e684bad84d184e138adf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
