+++ DeathRun/Assets/Scripts/Obs_scripts/IObstacle.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IObstacle
+{
+    void Activate(); // 모든 장애물이 구현할 공통 메서드
+}
