+++ DeathRun/Assets/Scripts/Obs_scripts/IObstacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b632dd47bd21df4a978e455f23fceed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
